Create storage engine dependent test cases
This adds tests that depend on the storage backend used. It
currently yields errors because table name excapes are different
in MySQL and SQLite. This will be developped to more complete
tests as soon as it is running on different backends correctly.

TODO: Fix escape bugs in AdoDBRecord
TODO: Complete basic tests