Fix classes to work with SQLite and make tests run
This fixes the classes to work with SQLite. This is a
quickfix: table and column names should be quoted
correctly later.

Also it enables #create() to create multiple objects at
once and the method was moved to dynamic implementation
classes to work around PHP's strange get_class()
behaviour.

"LIMIT" parameter was removed from #find because AdoDB
takes care of it by itself. #find("first") was fixed
to really work.

Hail to the SimpleTest framework for elaborating all
these bugs.

TODO Quote table and column names