Create inflectors needed for associations
This creates inflector methods to derive the class name from a
camel_cased name which is needed for creating the association options.

Also added an init.php for easier loading of the library. It sets
some defaults and changes the include path correctly.

Using the new inflector methods the default options of associations
are now set. Next step is doing the wrappers (marked relevant code
parts with TODO).